defaultpreemption: refactor out support for out-of-tree plugins
