+++ pkg/scheduler/framework/plugins/helper/candidate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package defaultpreemption
+package helper
 
 import (
 	extenderv1 "k8s.io/kube-scheduler/extender/v1"
